Create launchAppv5App.vbs
This script allows you to do per-user publishing.  It's targeted at XenApp 6.5 and 7.x, but should work with any environment where a user is presented with seamless applications.

The script tries to make sure that the application is published to the user.  If it's not, it will initiate a publishing refresh.  With Mid Sync Launch enabled, as soon as the application is available, the script will launch it.  There is logic built into the script to prevent multiple App-V Publishing Refreshes from happening at the same time (consider launching an application in a XenApp environment where sync on logon is enabled).

You need to customize the following variables:
  appName - displayed on the loading screen
  packageGUID - package GUID of the application you're launching.  Used to check for publishing rights.
  versionGUID - version GUID of the application you're launching.
  exeToLaunch - The full path to the EXE as it exists on a machine after it's published
  exeWorkingDir - Where you want to execute the EXE from
  exeParameters - Any Parameters you want passed to the executable at launch time


You can customize these:
Debug mode can be turned on by setting bDebug to true.

midSyncLaunch defaults to true.  It allows an application to launch in the as soon as the script detects that it's published to the user.  It speeds up launching considerably.  If you want to wait until all publishing is done before you launch this app, set midSyncLaunch to false.